@@ -1,13 +1,23 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-
 buildscript {
-    ext.kotlin_version = '1.1.51'
+    ext {
+        compileSdkVersion = 29
+        minSdkVersion = 16
+        targetSdkVersion = 29
+        kotlin_version = '1.3.60'
+        appcompat_version = '1.1.0'
+        junit_junit_version = '4.12'
+        androidx_test_ext_junit_version = '1.1.1'
+        espresso_core_version = '3.2.0'
+        constraintlayout_version = '1.1.3'
+        core_ktx_version = '1.1.0'
+        material_version = '1.0.0'
+    }
     repositories {
         jcenter()
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath 'com.android.tools.build:gradle:3.5.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
         // NOTE: Do not place your application dependencies here; they belong
