@@ -1,6 +1,6 @@
 package com.luoyang.shake
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import com.luoyang.shake.utils.ShakeUtil
