@@ -2,7 +2,7 @@ package com.luoyang.notification.activity;
 
 import android.content.Context;
 import android.content.Intent;
-import android.databinding.DataBindingUtil;
+import androidx.databinding.DataBindingUtil;
 import android.util.Log;
 import android.view.View;
 
